+++ backend/crypto/app.py
@@ -6,4 +6,4 @@ app = Flask(__name__)
 
 
 
-from crypto import routes
+# from crypto import routes
